stock_predictor/requirements.txt
pandas
numpy
xgboost
scikit-learn
ta
streamlit
python-dotenv
finnhub-python
plotly
